excel2word/requirements.txt
streamlit
openpyxl
python-docx
